Equip and sell items from inventory
the player can use items but all do nothing, cooldown works, and you can equip items and sell items from the inventory. There's also a default item for all active slots (slots that aren't in the inventory) which will show if the active item is null, and will not return when trying to swap items.